Added logger and get high stakes and refactored session id to LONG

+++ betthread/src/main/java/com/sorin/betthread/session/Session.java
@@ -1,14 +1,14 @@
 package com.sorin.betthread.session;
 
 public class Session {
-	private final int sessionId;
+	private final long sessionId;
 	private final String sessionKey;
 	private final int customerId;
 	private final int timeoutMilis; // faster than working with Dates ?
 	
 	private long expirationDate;
 	
-	public Session(int sessionId, int customerId, int timeoutMilis) {
+	public Session(long sessionId, int customerId, int timeoutMilis) {
 		this.sessionId = sessionId;
 		this.sessionKey = getKeyFromId(sessionId);
 		this.customerId = customerId;
@@ -17,7 +17,7 @@ public class Session {
 		updateExpiration();
 	}
 
-	public int getSessionId() {
+	public long getSessionId() {
 		return sessionId;
 	}
 
@@ -42,15 +42,15 @@ public class Session {
 		return this.expirationDate >= System.currentTimeMillis();
 	}
 	
-	public static final int getIdFromKey(String key) throws InvalidSessionKeyException {
+	public static final long getIdFromKey(String key) throws InvalidSessionKeyException {
 		try {
-			return Integer.parseInt(key, 16);
+			return Long.parseLong(key, 16);
 		} catch (Exception e) {
 			throw new InvalidSessionKeyException("Key " + key + " could not be parsed to a valid session id", e);
 		}
 	}
 	
-	public static final String getKeyFromId(int id) {
-		return Integer.toHexString(id).toUpperCase();
+	public static final String getKeyFromId(long id) {
+		return Long.toHexString(id).toUpperCase();
 	}
 }
